exported social link in component to /components/
