+++ client/src/components/Footer/Subscribe/index.jsx
@@ -1,23 +1,14 @@
 import React from 'react'
 import { styled } from '@mui/material/styles'
-import { Grid, List, ListItem, Typography,
-	Link as LinkMui } from '@mui/material'
+import { Grid, List, ListItem, Typography, } from '@mui/material'
 
 import FormSubscribe from '../FormSubscribe'
-
-import TwitterIcon from '@mui/icons-material/Twitter'
-import FacebookIcon from '@mui/icons-material/Facebook'
-import InstagramIcon from '@mui/icons-material/Instagram'
+import SocialLinks from '../../SocialLInks'
 
 const StyledListItem = styled(ListItem)(() => ({
 	flexDirection: 'row',
 	flexWrap: 'wrap',
 	justifyContent: 'flex-start',
-}))
-
-const StyledList = styled(List)(() => ({
-	display: 'flex',
-	flexDirection: 'row',
 }))
 
 
@@ -43,36 +34,9 @@ const Subscribe = () => {
 				</StyledListItem>
 				<ListItem>
 					<Typography variant="menuBold">
-								Follow us
+						follow us
 					</Typography>
-					<StyledList 
-						disablePadding={true}
-					>
-						<ListItem>
-							<LinkMui
-								href="https://facebook.com" 
-								target="_blank"
-							>
-								<FacebookIcon />
-							</LinkMui>
-						</ListItem>
-						<ListItem>
-							<LinkMui
-								href="https://twitter.com" 
-								target="_blank"
-							>
-								<TwitterIcon />
-							</LinkMui>
-						</ListItem>
-						<ListItem>
-							<LinkMui
-								href="https://instagram.com" 
-								target="_blank"
-							>
-								<InstagramIcon />
-							</LinkMui>
-						</ListItem>
-					</StyledList>
+					<SocialLinks />
 				</ListItem>
 			</List>
 		</Grid>
