+++ client/src/utils/Theme.js
@@ -38,7 +38,6 @@ const theme = createTheme({
 			fontSize: '16px',
 			lineHeight: 'normal'
 		},
-
 		footerTextMedium: {
 			textTransform: 'capitalize',
 			color: '#373F41',
